Feat: 오토인덱스 소스파일 추가
